feat: 예약 삭제 시, 첫 번째 예약 대기 인원 자동 전환 기능 추가
순서
- 사용자 예약 삭제 시, 같은 정보로 예약 대기 존재하는지 탐색
- 예약 대기가 존재할 경우 첫 번째 예약 대기 정보 반환
- 첫 번째 예약 대기 이름을 예약으로 update
- 예약 대기 삭제

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -12,17 +12,22 @@ import roomescape.reservation.dao.ReservationDao;
 import roomescape.theme.dao.ThemeDao;
 import roomescape.time.ReservationTime;
 import roomescape.time.dao.TimeDao;
+import roomescape.waiting.ReservationWaiting;
+import roomescape.waiting.dao.ReservationWaitingDao;
 
 @Service
 public class ReservationService {
     private final ReservationDao reservationDao;
     private final ThemeDao themeDao;
     private final TimeDao timeDao;
+    private final ReservationWaitingDao reservationWaitingDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, TimeDao timeDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, TimeDao timeDao,
+                              ReservationWaitingDao reservationWaitingDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.timeDao = timeDao;
+        this.reservationWaitingDao = reservationWaitingDao;
     }
 
     public List<Reservation> findAll() {
@@ -79,6 +84,7 @@ public class ReservationService {
         reservationDao.deleteById(id);
     }
 
+    @Transactional
     public void deleteByIdIfNameMatches(Long id, String name) {
         Reservation originReservation = reservationDao.selectById(id)
                 .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_NOT_FOUND));
@@ -88,7 +94,24 @@ public class ReservationService {
         validateDateTime(originReservation.getDate(), originReservation.getTime(),
                 ErrorCode.CANNOT_DELETE_PAST_RESERVATION);
 
+        if (validateExistsWaiting(originReservation.getThemeId(), originReservation.getDate(), originReservation.getTime().getId())) {
+            autoApprove(id, originReservation.getThemeId(), originReservation.getDate(), originReservation.getTime().getId());
+            return;
+        }
+
         reservationDao.deleteById(id);
+    }
+
+    private void autoApprove(Long id, Long themeId, LocalDate date, Long timeId) {
+        ReservationWaiting firstWaiting = reservationWaitingDao.selectFirstWaitingByThemeIdAndDateAndTimeId(themeId, date, timeId);
+
+        reservationDao.updateNameByThemeIdAndDateAndTimeId(id, firstWaiting.getName(), themeId, date, timeId)
+                .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_NOT_FOUND));
+        reservationWaitingDao.deleteById(firstWaiting.getId());
+    }
+
+    private boolean validateExistsWaiting(Long themeId, LocalDate date, Long timeId) {
+        return reservationWaitingDao.existsByThemeIdAndDateAndTimeId(themeId, date, timeId);
     }
 
     private void validateReserved(Long timeId, ReservationTime reservedTime) {
